+++ src/main/java/displays/Chessboard.java
@@ -1,4 +1,4 @@
-package gui;
+package displays;
 
 import application.MoveValidator;
 import bots.ChessBot;
