+++ src/main/java/displays/StartupScreen.java
@@ -1,4 +1,4 @@
-package gui;
+package displays;
 
 import bots.ChessBot;
 import bots.RandomBot;
