+++ src/test/java/displays/ChessboardTest.java
@@ -1,8 +1,7 @@
-package gui;
+package displays;
 
 import bots.RandomBot;
 import enums.ColourEnum;
-import gui.Chessboard;
 import org.junit.Test;
 import static org.hamcrest.CoreMatchers.notNullValue;
 import static org.hamcrest.MatcherAssert.assertThat;
